+++ apps/server/src/app/services/performance-test.interface.ts
@@ -1,0 +1,8 @@
+export interface PerformanceTest {
+  name: string;
+  description: string;
+  baselineTime: number;
+  optimizedTime: number;
+  improvement: number;
+  passed: boolean;
+}
